+++ parsley.js
@@ -18,7 +18,7 @@
   var Validator = function ( options ) {
     /**
     * Error messages
-    * 
+    *
     * @property messages
     * @type {Object}
     */
@@ -59,7 +59,7 @@
 
     /**
     * Validator list. Built-in validators functions
-    * 
+    *
     * @property validators
     * @type {Object}
     */
@@ -601,7 +601,7 @@
     * @returns {String} val
     */
     , getVal: function () {
-      return this.$element.val();
+      return this.$element.data('value') || this.$element.val();
     }
 
     /**
@@ -629,7 +629,7 @@
         return true;
       }
 
-      this.validate( true );
+      this.validate( );
     }
 
     /**
@@ -683,7 +683,7 @@
         return this.valid;
       }
 
-      this.errorBubbling = 'undefined' !== typeof errorBubbling ? errorBubbling : true;
+      this.errorBubbling = 'undefined' !== typeof errorBubbling ? errorBubbling : this.options.showErrors;
 
       valid = this.applyValidators();
 
@@ -1205,7 +1205,7 @@
 
       this.$element.off( '.' + this.type ).removeData( this.type );
     }
-    
+
     /**
     * reset Parsley binded on the form and its fields
     *
@@ -1281,7 +1281,7 @@
 
   /**
   * Parsley plugin configuration
-  * 
+  *
   * @property $.fn.parsley.defaults
   * @type {Object}
   */
@@ -1298,6 +1298,7 @@
     , errorClass: 'parsley-error'               // Class name on each invalid input
     , errorMessage: false                       // Customize an unique error message showed if one constraint fails
     , validators: {}                            // Add your custom validators functions
+    , showErrors: true                          // Set to false if you don't want Parsley to display error messages
     , messages: {}                              // Add your own error messages here
 
     //some quite advanced configuration here..