+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Changelog
 
+**1.1.16-dev**
+
+  - added option to silence errors `data-show-errors="false"`
+  - added / modified some i18n localizations
+  - Override value by using data-value attribute
+
 **1.1.15 (current stable)**
 
   - **BC Break** removed `.parsley('isFieldValid');` in favor of `.parsley('isValid');`